move chat panel in to button dialog in arbiter page
